backend/requirements.txt
beautifulsoup4==4.10.0
fastapi==0.75.0
loguru==0.6.0
pydantic==1.9.0
python-dotenv==0.20.0
python-multipart==0.0.5
python-slugify==6.1.1
requests==2.27.1
sqlmodel==0.0.6
uvicorn[standard]==0.17.6
